Refactored action mode and fixed UI issues.
