+++ test/core/services/logger_service_test.dart
@@ -58,10 +58,6 @@ void main() {
     final printer = _RecordingPrinter();
     rootLogger.printer = printer;
     var traceFormatted = false;
-    traceLog(() {
-      traceFormatted = true;
-      return 'frequent-event';
-    });
     appLogger.debug('ordinary-debug');
     appLogger.error('event-error');
     await Future<void>.delayed(Duration.zero);
